+++ fixtures/transform-actual/unattachable-mix.ts
@@ -1,0 +1,16 @@
+//=> unknown
+
+//=> unknown
+
+// @dts-jest
+[];
+//=> []
+
+//=> unknown
+
+//=> unknown
+
+//=> unknown
+
+// @dts-jest
+[]; //=> []