+++ fixtures/transform-actual/unattachable.ts
@@ -1,0 +1,9 @@
+//=> unknown
+
+//=> unknown
+
+//=> unknown
+
+//=> unknown
+
+//=> unknown