+++ src/index.ts
@@ -1,2 +1,3 @@
 export {setup} from './setup';
 export {transform} from './transform';
+export {transform_actual} from './transform-actual';