+++ transform-actual.js
@@ -1,0 +1,1 @@
+exports.process = require('./lib/index').transform_actual;